backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic[email]
sqlalchemy
tensorflow
pillow
numpy
